+++ packages/dataviews/CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
-## 3.0.0 (2024-07-10)
+## 4.0.0 (2024-07-24)
 
+### Breaking Changes
+
+- `onSelectionChange` prop has been renamed to `onChangeSelection` and its argument has been updated to be a list of ids.
+- `setSelection` prop has been removed. Please use `onChangeSelection` instead.
+- `header` field property has been renamed to `label`.
+
+## 3.0.0 (2024-07-10)
 
 ### Breaking Changes